move BundleClassLoaderProvider class to internal.io.deployment package
